New Editor Windows Developed and the system was developed to save and load the data as json.
Create Symbol Editor: Editor window that allows new symbols to be added to the slot game.
Create Result Editor: Editor window that allows adding a new result to the slot game.
Show Results: Editor window that shows the player all upcoming spin results and which spin the player is currently on.
JsonSaver: A static class that allows data to be saved and loaded. Newtonsoft.json package was used.